+++ test-data/projects/deps-main-opts/src/main_opts/hello.clj
@@ -1,0 +1,4 @@
+(ns main-opts.hello)
+
+(defn -main [& args]
+  (println "Hello"))