+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.352] - 2023-04-16
+
 - [Add custom Jack-in command lines to connect sequences](https://github.com/BetterThanTomorrow/calva/issues/2152)
 - [Add docs about main-opts and make the warning link there](https://github.com/BetterThanTomorrow/calva/issues/2157)